check illegal sampling rates for AAC, MP4
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@3366 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/sidplay/audio/MP4FileDriver.java
@@ -38,9 +38,11 @@ import org.sheinbergon.aac.encoder.util.AACEncodingProfile;
 import libsidplay.common.CPUClock;
 import libsidplay.common.EventScheduler;
 import libsidplay.common.OS;
+import libsidplay.common.SamplingRate;
 import libsidplay.components.mos656x.VIC;
 import libsidplay.config.IAudioSection;
 import libsidutils.PathUtils;
+import sidplay.audio.exceptions.IniConfigException;
 
 public class MP4FileDriver implements AudioDriver, VideoDriver {
 
@@ -62,6 +64,11 @@ public class MP4FileDriver implements AudioDriver, VideoDriver {
 		AudioConfig cfg = new AudioConfig(audioSection);
 		this.recordingFilename = recordingFilename;
 		try {
+			if (audioSection.getSamplingRate().getFrequency() == 8000
+					|| audioSection.getSamplingRate().getFrequency() == 96000) {
+				throw new IniConfigException("Sampling rate is not supported by AACEncoder",
+						() -> audioSection.setSamplingRate(SamplingRate.MEDIUM));
+			}
 			System.out.println("Recording, file=" + recordingFilename);
 			new File(recordingFilename).delete();
 			String recordingBaseName = PathUtils.getFilenameWithoutSuffix(recordingFilename);
